LayoutTests/imported/w3c/web-platform-tests/content-security-policy/style-src/stylenonce-blocked.sub-expected.txt

This text should be green.

Style that does not match a 'nonce-*' expression in 'style-src' should not be applied to the page.


FAIL Should fire securitypolicyviolation assert_equals: expected "style-src-elem" but got "style-src"
PASS stylenonce-blocked
